backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.4",
    "google-adk>=1.18.0",
    "orjson>=3.11.5",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.27",
    "supabase>=2.25.1",
    "uvicorn[standard]>=0.38.0",
    "numpy>=1.26.0",
]

[dependency-groups]
dev = [
    "bandit>=1.9.2",
    "black>=25.12.0",
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.7.0",
    "mkdocstrings[python]>=1.0.0",
    "mypy>=1.19.0",
    "pip-audit>=2.10.0",
    "pre-commit>=4.5.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]
